Implement adaptive frame delay
If the sync plugin detects that the game is out of sync (i.e. we missed
a frame), then we will skip delay on the next frame to try to get back
in sync.

This makes it possible to push frame delay closer to the limit with
less risk of video stutter (at the expense of some jitter in when input
is polled).