Add option to delay swap with a query object
Added an option to delay swap buffers using a query object.  This is
like glFinish, but with a deadline.  This is an improvement over
nanosleep, because if the query object is available, rendering is
complete and we can swap buffers right away, whereas sleeping for the
entire delay duration risks overshooting the deadline.